Change SAVE_FORMAT_VERSION to a simple unsigned int

+++ rerelease/g_save.cpp
@@ -22,7 +22,7 @@
 // - backwards & forwards compatible with this same format
 // - I wrote this initially when the codebase was in C, so it
 //   does have some C-isms in here.
-constexpr size_t SAVE_FORMAT_VERSION = 1;
+constexpr unsigned int SAVE_FORMAT_VERSION = 1;
 
 #include <unordered_map>
